Cast runtime differences to ints
Casting the differences to ints keeps the telemetry on one line.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/GoldContourPipelineTest.java
@@ -121,10 +121,10 @@ public class GoldContourPipelineTest extends LinearOpMode {
             telemetry.addData("Hue max", hsvHue[1]);
             telemetry.addLine();
             telemetry.addData("dpUpReady", dpUpReady);
-            telemetry.addData("dpUpRuntime - snapshot", dpUpRuntimeDif);
+            telemetry.addData("dpUpRuntime - snapshot", (int)dpUpRuntimeDif);
             telemetry.addLine();
             telemetry.addData("dpDownReady", dpDownReady);
-            telemetry.addData("dpDownRuntime - snapshot", dpDownRuntimeDif);
+            telemetry.addData("dpDownRuntime - snapshot", (int)dpDownRuntimeDif);
             telemetry.update();
         }
